Implement strategy pattern to supply game grid.
Overkill for now, but my plan is to eventually have a single linear array for the game grid.
But for now, I want to use a simpler 2D array.
So by going with a strategy pattern like this, I'm setting myself up to be able to swap in the linear grid later, once I finish core development with the easier to visualise 2D grid.