Encapsulate coordinates, height and width.
I decided to encapsulate the coordinates so that I could use varargs in the parameterised tests. A pair would have sufficed, but then you need to know which is x and which is y, so a record to hold them seemed best.
The Height and Width being objects is probably overkill really. I just felt uncomfortable about the possibility of mixing them up when instantiating the grid.