Add complete implementation of taking a turn.
This involved some fairly substantial reworking. I had hoped to be able to keep the game grid consisting of just the boolean state, but then I was struggling to get a clean way to implement the logic I needed.

I was running into an issue when testing the output of a turn, in that I could not match the order of output specified in the brief.

For now, I added a hacky "sort," which means the test is only validating the correct contents, and not the correct order.